modularize slice ops
