+++ tfjs-core/src/ops/ops.ts
@@ -116,6 +116,11 @@ export {reverse4d} from './reverse_4d';
 export {selu} from './selu';
 export {separableConv2d} from './separable_conv2d';
 export {sign} from './sign';
+export {slice} from './slice';
+export {slice1d} from './slice1d';
+export {slice2d} from './slice2d';
+export {slice3d} from './slice3d';
+export {slice4d} from './slice4d';
 export {spaceToBatchND} from './space_to_batch_nd';
 export {split} from './split';
 export {square} from './square';
@@ -132,7 +137,6 @@ export {where} from './where';
 export {whereAsync} from './where_async';
 
 export * from './boolean_mask';
-export * from './slice';
 export * from './unary_ops';
 export * from './compare';
 export * from './binary_ops';
