+++ tfjs-core/src/ops/slice4d.ts
@@ -1,0 +1,40 @@
+/**
+ * @license
+ * Copyright 2018 Google LLC. All Rights Reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =============================================================================
+ */
+
+import {Tensor4D} from '../tensor';
+import {convertToTensor} from '../tensor_util_env';
+import {TensorLike} from '../types';
+import * as util from '../util';
+
+import {op} from './operation';
+import {slice} from './slice';
+
+/**
+ * Extracts a 4D slice from a 4D array starting at coordinates `begin` and
+ * is of size `size`. See `slice` for details.
+ */
+function slice4d_(
+    x: Tensor4D|TensorLike, begin: [number, number, number, number],
+    size: [number, number, number, number]): Tensor4D {
+  const $x = convertToTensor(x, 'x', 'slice4d');
+  util.assert(
+      $x.rank === 4,
+      () =>
+          `slice4d expects a rank-4 tensor, but got a rank-${$x.rank} tensor`);
+  return slice($x, begin, size);
+}
+export const slice4d = op({slice4d_});
