+++ tfjs-core/src/tests.ts
@@ -147,6 +147,10 @@ import './ops/selu_test';
 import './ops/sigmoid_cross_entropy_test';
 import './ops/sign_test';
 import './ops/signal_ops_test';
+import './ops/slice1d_test';
+import './ops/slice2d_test';
+import './ops/slice3d_test';
+import './ops/slice4d_test';
 import './ops/slice_test';
 import './ops/slice_util_test';
 import './ops/softmax_cross_entropy_test';
